Actualización de scripts SQL y creación de scripts de popular y extraer

+++ BaseDeDatos/data/crear_tabla_datos_meteorologicos.sql
@@ -3,7 +3,7 @@ CREATE TABLE datos_meteorologicos (
     indicativo_estacion VARCHAR(10) NOT NULL,
     nombre_estacion VARCHAR(100),
     -- nombre_provincia VARCHAR(50), -- No la guardamos si ya tenemos codigo_provincia
-    codigo_provincia SMALLINT UNSIGNED, -- FK a provincias
+    codigo_provincia TINYINT UNSIGNED, -- FK a provincias
     altitud SMALLINT UNSIGNED,
     temperatura_media DECIMAL(4,1),
     precipitacion DECIMAL(5,1),
